+++ src/App.tsx
@@ -10,7 +10,7 @@ import {
 import { useAuth } from "./auth/AuthProvider";
 import UserDashboard from "./pages/Dashboard/UserDashboard";
 import AdminDashboard from "./pages/Dashboard/AdminDashboard";
-import AdminCreator from "./pages/Dashboard/AdminCreator";
+import AdminCreator from "./pages/Users/AdminCreator";
 import MatchedDashboard from "./pages/Dashboard/MatchedDashboard";
 import LoginSignup from "./pages/Login/Login-Signup";
 import Registration from "./pages/Registration/RegistrationNew";
